requirements.txt
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.11.1
better-exceptions==0.3.3
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
colored==1.3.93
cryptography==39.0.0
cssselect==1.2.0
DBUtils==3.0.2
decorator==5.1.1
dnspython==2.2.1
exceptiongroup==1.1.0
executing==1.2.0
feapder==1.8.5
greenlet==2.0.1
h11==0.14.0
idna==3.4
influxdb==5.3.1
iniconfig==2.0.0
ipython==8.8.0
jedi==0.18.2
loguru==0.6.0
lxml==4.9.2
matplotlib-inline==0.1.6
msgpack==1.0.4
outcome==1.2.0
packaging==23.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
playwright==1.29.1
pluggy==1.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyee==9.0.4
Pygments==2.14.0
pymongo==4.3.3
PyMySQL==1.0.2
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.2.0
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.7
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminal-layout==2.1.6
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==1.5.3
